+++ telemetry-test/samizdat/telemetry/test_runner.clj
@@ -22,7 +22,9 @@
   (:require [clojure.test :as t]
             [samizdat.telemetry.aspect-manifest-test]
             [samizdat.telemetry.contract-test]
+            [samizdat.telemetry.embedded-bootstrap-test]
             [samizdat.telemetry.embedded-test]
+            [samizdat.telemetry.embedded-serve-test]
             [samizdat.telemetry.hook-test]
             [samizdat.telemetry.otel-test]
             [samizdat.telemetry.pipelines-test]))
@@ -30,7 +32,9 @@
 (def namespaces
   '[samizdat.telemetry.aspect-manifest-test
     samizdat.telemetry.contract-test
+    samizdat.telemetry.embedded-bootstrap-test
     samizdat.telemetry.embedded-test
+    samizdat.telemetry.embedded-serve-test
     samizdat.telemetry.hook-test
     samizdat.telemetry.otel-test
     samizdat.telemetry.pipelines-test])